# backend/public/api-docs/openapi.yaml
openapi: 3.0.0
info:
  title: Sassy CRM API
  description: Modern headless CRM API for software sales teams with AI-powered features
  version: 1.0.0
  contact:
    name: API Support
    email: support@sassycrm.com
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.sassycrm.com/api
    description: Production server
security:
  - bearerAuth: []
...
paths:
...
  /crm/leads:
    get:
      tags: [Leads]
      summary: List leads
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: status
          in: query
          schema:
            type: string
            enum: [new, contacted, qualified, unqualified, converted]
        - name: assigned_user_id
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        200:
          description: List of leads
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Lead'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags: [Leads]
      summary: Create lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeadInput'
      responses:
        201:
          description: Lead created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                  message:
                    type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
...
    Lead:
      type: object
      properties:
        id:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        email1:
          type: string
        email2:
          type: string
        phone_work:
          type: string
        phone_mobile:
          type: string
        phone_home:
          type: string
        phone_other:
          type: string
        phone_fax:
          type: string
        title:
          type: string
        department:
          type: string
        account_name:
          type: string
        website:
          type: string
        lead_source:
          type: string
        lead_source_description:
          type: string
        status:
          type: string
          enum: [new, contacted, qualified, unqualified, converted]
        status_description:
          type: string
        description:
          type: string
        assigned_user_id:
          type: string
        date_entered:
          type: string
          format: date-time
        date_modified:
          type: string
          format: date-time
        ai_score:
          type: number
          format: float
    LeadInput:
      type: object
      required: [last_name]
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email1:
          type: string
        email2:
          type: string
        phone_work:
          type: string
        phone_mobile:
          type: string
        phone_home:
          type: string
        phone_other:
          type: string
        phone_fax:
          type: string
        title:
          type: string
        department:
          type: string
        account_name:
          type: string
        website:
          type: string
        lead_source:
          type: string
        lead_source_description:
          type: string
        status:
          type: string
          enum: [new, contacted, qualified, unqualified, converted]
        status_description:
          type: string
        description:
          type: string
        assigned_user_id:
          type: string
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer